+++ app/src/main/java/org/nervos/neuron/fragment/WalletsFragment/presenter/WalletFragmentPresenter.java
@@ -7,9 +7,6 @@ import android.util.TypedValue;
 import android.view.View;
 import android.widget.LinearLayout;
 
-import org.nervos.neuron.R;
-import org.nervos.neuron.util.db.SharePrefUtil;
-
 import java.lang.reflect.Field;
 
 /**
@@ -21,19 +18,6 @@ public class WalletFragmentPresenter {
 
     public WalletFragmentPresenter(Activity activity) {
         this.activity = activity;
-    }
-
-    public String getTotalMoneyTitle() {
-        String str = activity.getResources().getString(R.string.wallet_token_unit_CNY);
-        switch (SharePrefUtil.getString("Currency", "CNY")) {
-            case "CNY":
-                str = activity.getResources().getString(R.string.wallet_token_unit_CNY);
-                break;
-            case "USD":
-                str = activity.getResources().getString(R.string.wallet_token_unit_USD);
-                break;
-        }
-        return "（" + str + "）";
     }
 
     public void setIndicator(TabLayout tabs, int leftDip, int rightDip) {
