+++ app/src/main/res/layout/fragment_wallets.xml
@@ -72,49 +72,6 @@
             android:layout_height="1.5dp"
             android:background="@color/line_form" />
 
-        <RelativeLayout
-            android:id="@+id/ll_total_money"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:background="@color/white">
-
-            <TextView
-                android:id="@+id/tv_total_money_title"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_centerVertical="true"
-                android:layout_marginLeft="20dp"
-                android:text="@string/wallet_token_total_money"
-                android:textColor="@color/font_title_second"
-                android:textSize="14sp" />
-
-            <TextView
-                android:id="@+id/tv_total_money"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_centerVertical="true"
-                android:layout_marginLeft="2dp"
-                android:layout_toRightOf="@id/tv_total_money_title"
-                android:textColor="@color/font_title_second"
-                android:textSize="14sp" />
-
-            <ImageView
-                android:id="@+id/iv_add"
-                android:layout_width="25dp"
-                android:layout_height="25dp"
-                android:layout_alignParentRight="true"
-                android:layout_marginBottom="10dip"
-                android:layout_marginRight="20dp"
-                android:layout_marginTop="10dip"
-                android:src="@drawable/ic_wallet_token_add" />
-
-            <View
-                android:layout_width="match_parent"
-                android:layout_height="1.5dp"
-                android:layout_alignParentBottom="true"
-                android:background="@color/line_form" />
-        </RelativeLayout>
-
     </android.support.design.widget.AppBarLayout>
 
     <android.support.v4.widget.NestedScrollView
@@ -127,7 +84,6 @@
             android:id="@+id/viewPager"
             android:layout_width="match_parent"
             android:layout_height="match_parent" />
-
     </android.support.v4.widget.NestedScrollView>
 
 </android.support.design.widget.CoordinatorLayout>
